requirements.txt
clickhouse-connect
scrapy
google-generativeai
beautifulsoup4
python-dotenv
PyMuPDF
requests
psycopg2-binary
